chore: break up circle CI tasks
