+++ package.json
@@ -18,6 +18,7 @@
     "build-template-html-watch": "node -r esm ./bin/build-template-html.js --watch",
     "build-assets": "node -r esm ./bin/build-assets.js",
     "clone-mastodon": "node -r esm ./bin/clone-mastodon.js",
+    "install-mastodon": "node -r esm ./bin/install-mastodon.js",
     "run-mastodon": "node -r esm ./bin/run-mastodon.js",
     "test": "cross-env BROWSER=chrome:headless run-s test-browser",
     "test-browser": "run-p --race run-mastodon build-and-start test-mastodon",
@@ -26,6 +27,7 @@
     "testcafe-suite0": "cross-env-shell testcafe -c 4 $BROWSER tests/spec/0*",
     "testcafe-suite1": "cross-env-shell testcafe $BROWSER tests/spec/1*",
     "test-unit": "NODE_ENV=test mocha -r esm -r bin/browser-shim.js tests/unit/",
+    "test-in-ci": "cross-env BROWSER=chrome:headless run-p --race run-mastodon start test-mastodon",
     "wait-for-mastodon-to-start": "node -r esm bin/wait-for-mastodon-to-start.js",
     "wait-for-mastodon-data": "node -r esm bin/wait-for-mastodon-data.js",
     "backup-mastodon-data": "./bin/backup-mastodon-data.sh",
